+++ firmware/src/ble_service.h
@@ -9,3 +9,4 @@ const char* ble_service_take_data();
 void ble_service_ack();
 void ble_service_nack();
 void ble_service_request_refresh();
+bool ble_service_connected();